the first draft of sailor consumer client
consumer is stateful client which manages resources for your cloud native applications. This can be specificially viewed as the client that manages your sailor resources for backend application consumption.

NOTE: this is not a rest client, soon a rest client will be created which will expose all the APIs of Sailor.