feat: sailor consumer now is created with generics
sailor consumer is designed to be used in core backend services where sailor is responsible to hold and update configs and secrets for the app, in order to make it more friendly to use we are now going to store atomic.Pointer to the type C and S which are generic type parameters to Consumer[C, S]. This consumer can then be used for the whole lifecycle of the app.

The ergonomic way is to inject the `Consumer` as a dependency and call `consumer.Get` for usage in a single thread op.